[FEAT]#7 지원서 로직 구현 - api 없는 버전
