+++ src/component/Application/Textarea.tsx
@@ -18,19 +18,15 @@ const Wrapper = styled.div`
 const StyledTextarea = styled.textarea`
   width: 32.375rem;
   height: 10rem;
-  flex-shrink: 0;
   border-radius: 0.625rem;
   border: 1px solid #dbdbdf;
   background: #fafafa;
 
   color: #000;
-  leading-trim: both;
-  text-edge: cap;
-  font-family: Gmarket Sans TTF;
   font-size: 0.875rem;
   font-style: normal;
   font-weight: 300;
-  line-height: normal;
+  line-height: 1.3rem;
 `;
 
 const Textarea: React.FC<TextareaProps> = ({ name, value, onChange }) => {
