+++ src/index.tsx
@@ -1,12 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-
-import App from './App';
+import ReactDOM from 'react-dom/client';
 import './index.css';
+import App from './App';
 
-ReactDOM.render(
+const rootElement = document.getElementById('root');
+if (!rootElement) throw new Error('Failed to find the root element');
+const root = ReactDOM.createRoot(rootElement);
+
+root.render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 );
